Vehicle class updates
added setters for direction, velocity and position. Added a move function

+++ SSDCTN_traffic_sim/src/Velocity.java
@@ -1,7 +1,7 @@
 //import java.math.*;
 
 public class Velocity {
-    static public Point calPosition(double x, double y, float direction, float velocity){
+    static public Point calPosition(double x, double y, double direction, float velocity){
         double xPosAddition = Math.cos(direction) * velocity;
         double yPosAddition = Math.sin(direction) * velocity;
         x += xPosAddition;
